faults depth fields to earthquakes
