+++ src/org/opensha2/eq/model/InterfaceSource.java
@@ -3,8 +3,8 @@ package org.opensha2.eq.model;
 import static com.google.common.base.Preconditions.checkNotNull;
 import static com.google.common.base.StandardSystemProperty.LINE_SEPARATOR;
 
-import static org.opensha2.eq.fault.Faults.validateInterfaceDepth;
-import static org.opensha2.eq.fault.Faults.validateInterfaceWidth;
+import static org.opensha2.eq.Earthquakes.validateInterfaceDepth;
+import static org.opensha2.eq.Earthquakes.validateInterfaceWidth;
 import static org.opensha2.eq.fault.Faults.validateTrace;
 import static org.opensha2.eq.model.SourceType.INTERFACE;
