+++ src/org/opensha2/gmm/AbrahamsonEtAl_2014.java
@@ -17,6 +17,7 @@ import static org.opensha2.gmm.GmmInput.Field.Z1P0;
 import static org.opensha2.gmm.GmmInput.Field.ZTOP;
 
 import org.opensha2.data.Interpolate;
+import org.opensha2.eq.Earthquakes;
 import org.opensha2.eq.fault.Faults;
 import org.opensha2.gmm.GmmInput.Constraints;
 import org.opensha2.util.Maths;
@@ -54,8 +55,8 @@ public final class AbrahamsonEtAl_2014 implements GroundMotionModel {
       .set(MW, Range.closed(3.0, 8.5))
       .setDistances(300.0)
       .set(DIP, Faults.DIP_RANGE)
-      .set(WIDTH, Faults.CRUSTAL_WIDTH_RANGE)
-      .set(ZTOP, Faults.CRUSTAL_DEPTH_RANGE)
+      .set(WIDTH, Earthquakes.CRUSTAL_WIDTH_RANGE)
+      .set(ZTOP, Earthquakes.CRUSTAL_DEPTH_RANGE)
       .set(RAKE, Faults.RAKE_RANGE)
       .set(VS30, Range.closedOpen(180.0, 1000.0))
       .set(VSINF)
